+++ src/main/java/com/celfocus/training/entity/ShoppingCart.java
@@ -6,7 +6,7 @@ public class ShoppingCart {
     public User user;
     public List<ShoppingCartItem> shoppingCartItems;
 
-    public ShoppingCartItem getItemInShoppingCartFromName(String itemName) {
+    public ShoppingCartItem getShoppingCartItemFromItemName(String itemName) {
         return this.shoppingCartItems.stream().filter(itens -> itens.itemInfo.itemName.equals(itemName)).findAny().orElse(null);
     }
 }
