+++ src/main/java/com/celfocus/training/interfaces/ItemInfoInterface.java
@@ -1,0 +1,10 @@
+package com.celfocus.training.interfaces;
+
+import com.celfocus.training.entity.ItemInfo;
+
+import java.util.List;
+
+public interface ItemInfoInterface {
+    void addItemIfNotExists(String itemName, double itemValue);
+    ItemInfo getItemInfoFromName(String itemName);
+}
